refactor: encapsulate view structure handling on SceneHirarchyController #6
