+++ OrderApp/SceneDelegate.swift
@@ -6,21 +6,19 @@
 //
 
 import UIKit
-import Combine
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
     
-    private var orderUpdateSubscribe: Cancellable?
-    private weak var orderTabBarItem: UITabBarItem?
+    private var sceneHirarchyController = SceneHirarchyController()
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let _ = (scene as? UIWindowScene) else { return }
-        configureTabBarUI()
+        sceneHirarchyController.configure(with: self)
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
@@ -28,8 +26,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This occurs shortly after the scene enters the background, or when its session is discarded.
         // Release any resources associated with this scene that can be re-created the next time the scene connects.
         // The scene may re-connect later, as its session was not necessarily discarded (see `application:didDiscardSceneSessions` instead).
-        orderUpdateSubscribe?.cancel()
-        
     }
 
     func sceneDidBecomeActive(_ scene: UIScene) {
@@ -56,23 +52,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 }
 
 // MARK: RootViewController Handling Code
-extension SceneDelegate {
+extension SceneDelegate: SceneHirarchyControllerDelegate {
     
-    private func configureTabBarUI() {
-        guard let rootTabBarController = window?.rootViewController as? UITabBarController,
-              let orderTabBarItem = rootTabBarController.viewControllers?[1].tabBarItem else {
-            return
-        }
-        
-        self.orderTabBarItem = orderTabBarItem
-        
-        orderUpdateSubscribe = NotificationCenter.default.publisher(
-            for: .orderUpdateNotification,
-            object: nil
-        ).sink(receiveValue: { [weak self] _ in
-            let badgeValue = RestaurantController.shared.order.menuItems.count
-            self?.orderTabBarItem?.badgeValue = (badgeValue == 0) ? nil : String(badgeValue)
-        })
+    func loadUIHirarchy() -> UIWindow? {
+        return self.window
     }
     
 }
