+++ OrderApp/SceneHirarchyController.swift
@@ -1,0 +1,54 @@
+//
+//  SceneHirarchyController.swift
+//  OrderApp
+//
+//  Created by JeongTaek Han on 2023-02-04.
+//
+
+import UIKit
+import Combine
+
+final class SceneHirarchyController {
+    
+    weak var delegate: SceneHirarchyControllerDelegate?
+    
+    private weak var window: UIWindow?
+    
+    private var orderUpdateSubscribe: Cancellable?
+    private weak var orderTabBarItem: UITabBarItem?
+    
+    func configure(with delegate: SceneHirarchyControllerDelegate) {
+        self.delegate = delegate
+        window = self.delegate?.loadUIHirarchy()
+        configureTabBarUI()
+        subscribe()
+    }
+    
+    private func configureTabBarUI() {
+        guard let rootTabBarController = window?.rootViewController as? UITabBarController,
+              let orderTabBarItem = rootTabBarController.viewControllers?[1].tabBarItem else {
+            return
+        }
+        
+        self.orderTabBarItem = orderTabBarItem
+    }
+    
+    private func subscribe() {
+        orderUpdateSubscribe = NotificationCenter.default.publisher(
+            for: .orderUpdateNotification,
+            object: nil
+        ).sink(receiveValue: { [weak self] _ in
+            let badgeValue = RestaurantController.shared.order.menuItems.count
+            self?.orderTabBarItem?.badgeValue = (badgeValue == 0) ? nil : String(badgeValue)
+        })
+    }
+    
+    private func unsubscribe() {
+        orderUpdateSubscribe?.cancel()
+    }
+    
+    deinit {
+        unsubscribe()
+    }
+    
+}
